+++ src/pages/details/style.css.ts
@@ -62,3 +62,29 @@ export const fadeOut = style({
   animationName: fadeOutAnimation,
   animationFillMode: "forwards",
 });
+
+export const virtuosoContainer = style({
+  height: "auto",
+  width: "100%",
+  scrollbarWidth: "none",
+  "::-webkit-scrollbar": {
+    display: "none",
+  },
+});
+
+export const claimRewardsContainer = style([
+  atoms({
+    background: "positionsClaimRewardsBackground",
+    borderRadius: "base",
+  }),
+  {
+    padding: "2px 6px",
+  },
+]);
+
+export const viaText = style({
+  textOverflow: "ellipsis",
+  overflow: "hidden",
+  whiteSpace: "nowrap",
+  width: "100%",
+});
