+++ src/pages/steps/stake-steps.page.tsx
@@ -1,0 +1,8 @@
+import { useStakeState } from "../../state/stake";
+import { StepsPage } from "./common.page";
+
+export const StakeStepsPage = () => {
+  const { stakeSession } = useStakeState();
+
+  return <StepsPage session={stakeSession} />;
+};
